% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{new_location}
\alias{new_location}
\title{Create a single Location object (S3)}
\usage{
new_location(name, lat, lng)
}
\arguments{
\item{name}{Character. Human-readable place name.}

\item{lat}{Numeric. Latitude in decimal degrees (-90 to 90).}

\item{lng}{Numeric. Longitude in decimal degrees (-180 to 180).}
}
\value{
Object of class \code{tsp_location}.
}
\description{
Create a single Location object (S3)
}
\examples{
loc <- new_location("Warsaw", 52.23, 21.01)
print(loc)
}
