% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{new_tsp_route}
\alias{new_tsp_route}
\title{Create a TSP route solution object (S3)}
\usage{
new_tsp_route(tour, loc_set, dist_mat, algorithm, elapsed)
}
\arguments{
\item{tour}{Integer vector: city visit order (1-indexed).}

\item{loc_set}{A \code{tsp_location_set} object.}

\item{dist_mat}{The distance matrix used to solve.}

\item{algorithm}{Character. Name of the algorithm used.}

\item{elapsed}{Numeric. Computation time in seconds.}
}
\value{
Object of class \code{tsp_route}.
}
\description{
Create a TSP route solution object (S3)
}
\keyword{internal}
