% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{plot.tsp_route}
\alias{plot.tsp_route}
\title{Plot a TSP route}
\usage{
\method{plot}{tsp_route}(x, ...)
}
\arguments{
\item{x}{A \code{tsp_route} object.}

\item{...}{Passed to \code{plot()}.}
}
\value{
\code{x} invisibly.
}
\description{
Plot a TSP route
}
