% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{print.tsp_route}
\alias{print.tsp_route}
\title{Print a TSP route solution}
\usage{
\method{print}{tsp_route}(x, ...)
}
\arguments{
\item{x}{A \code{tsp_route} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Shows the algorithm used, total distance, computation time,
and the city visit order.
}
