% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{TSPSolver}
\alias{TSPSolver}
\title{TSP Solver}
\description{
R6 class that solves the Traveling Salesman Problem using the
Nearest Neighbour heuristic followed by 2-opt local search improvement.
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{n_locations}}{Number of loaded locations (read-only).}

    \item{\code{locations}}{The loaded LocationSet (read-only).}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TSPSolver-initialize}{\code{TSPSolver$new()}}
    \item \href{#method-TSPSolver-solve}{\code{TSPSolver$solve()}}
    \item \href{#method-TSPSolver-print}{\code{TSPSolver$print()}}
    \item \href{#method-TSPSolver-clone}{\code{TSPSolver$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TSPSolver-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TSPSolver-initialize}{}}}
\subsection{\code{TSPSolver$new()}}{
  Load locations and precompute distance matrix.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TSPSolver$new(df)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{df}}{Data frame with columns: name, lat, lng.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TSPSolver-solve"></a>}}
\if{latex}{\out{\hypertarget{method-TSPSolver-solve}{}}}
\subsection{\code{TSPSolver$solve()}}{
  Solve using Nearest Neighbour + 2-opt improvement.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TSPSolver$solve()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{tsp_route} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TSPSolver-print"></a>}}
\if{latex}{\out{\hypertarget{method-TSPSolver-print}{}}}
\subsection{\code{TSPSolver$print()}}{
  Print solver state.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TSPSolver$print(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TSPSolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TSPSolver-clone}{}}}
\subsection{\code{TSPSolver$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TSPSolver$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
